move basic_rgb
